+++ Mundi/Source/Runtime/Renderer/FViewportClient.cpp
@@ -188,7 +188,7 @@ void FViewportClient::MouseButtonDown(FViewport* Viewport, int32 X, int32 Y, int
         }
         Camera->SetWorld(World);
         UE_LOG("%f %f", ViewportMousePos.X, ViewportMousePos.Y);
-        PickedComponent = URenderManager::GetInstance().GetRenderer()->GetPrimitiveCollided(ViewportMousePos.X, ViewportMousePos.Y);
+        PickedComponent = URenderManager::GetInstance().GetRenderer()->GetPrimitiveCollided(static_cast<int>(ViewportMousePos.X), static_cast<int>(ViewportMousePos.Y));
        // PickedActor = CPickingSystem::PerformViewportPicking(AllActors, Camera, ViewportMousePos, ViewportSize, ViewportOffset, PickingAspectRatio,  Viewport);
